Create Titanic: Machine Learning from Disaster
In this data cleaning and standardisation project, the Titanic dataset was worked on to demonstrate skills in preparing data for analysis and modelling. The Titanic dataset, obtained from Kaggle, provided information about the passengers aboard the Titanic.

To begin, the dataset was loaded using Python and the Pandas library. Through exploratory data analysis, insights were gained by examining the dataset's structure and the initial rows. This allowed a better understanding of the available features and their data types.

Next, missing values in the dataset were handled. Missing age values were filled with the median age to ensure a more complete and representative dataset. Similarly, missing embarked values were filled with the mode, representing the most frequent value for the port of embarkation. This ensured the integrity and accuracy of the dataset.

To further enhance the dataset, data types were corrected. For instance, the fare column was converted to float, and the age column was converted to integer, providing consistent and appropriate data representations.

Additionally, outliers in the dataset were addressed to prevent them from significantly impacting analysis outcomes. Fare outliers were removed, considering values within three standard deviations from the mean. This ensured that extreme values did not skew the overall analysis.

Finally, the cleaned dataset was saved as a new CSV file for future use. This preserved the integrity of the original dataset while providing a clean and standardised version for analysis or modelling purposes.

By successfully completing this project, proficiency in handling real-world data, applying essential data cleaning techniques, and ensuring data quality was demonstrated. This project serves as a testament to the skills in data cleaning and transformation, highlighting attention to detail and the ability to prepare datasets for meaningful analysis or modelling tasks.